+++ packages/cli/src/__tests__/e2e/help.spec.ts
@@ -1,27 +1,25 @@
-import path from "path";
 import { clearStyle, w3Cli } from "./utils";
 
 import { runCLI } from "@web3api/test-env-js";
 
 const HELP = `
-  w3             🔥 Web3API CLI 🔥                                      
-  help (h)       -                                                    
-  test-env (t)   Manage a test environment for Web3API                
-  query (q)      Query Web3APIs using recipe scripts                  
-  plugin (p)     Build/generate types for the plugin                  
-  create (c)     Create a new project with w3 CLI                     
-  codegen (g)    Auto-generate API Types                              
-  build (b)      Builds a Web3API and (optionally) uploads it to IPFS 
-  app (a)        Build/generate types for your app                    
+  w3             🔥 Web3API CLI 🔥                       
+  help (h)       -                                     
+  test-env (t)   Manage a test environment for Web3API 
+  query (q)      Query Web3APIs using recipe scripts   
+  plugin (p)     Build/generate types for the plugin   
+  deploy (b)     Deploys/Publishes a Web3API           
+  create (c)     Create a new project with w3 CLI      
+  codegen (g)    Auto-generate API Types               
+  build (b)      Builds a Web3API                      
+  app (a)        Build/generate types for your app     
 `;
 
 describe("e2e tests for no help", () => {
-  const projectRoot = path.resolve(__dirname, "../project/");
 
   test("Should display the help content", async () => {
     const { exitCode: code, stdout: output, stderr: error } = await runCLI({
       args: ["help"],
-      cwd: projectRoot,
       cli: w3Cli,
     });