+++ packages/js/client/src/pluginConfigs/Ens.ts
@@ -1,8 +1,9 @@
 /// NOTE: This is an auto-generated file. See scripts/extractPluginConfigs.ts
 /* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable prettier/prettier */
 
 /// Types generated from @web3api/ens-plugin-js build files:
-/// build/w3-man/plugin.d.ts, build/query/index.d.ts
+/// build/w3/plugin.d.ts, build/query/index.d.ts
 
 export interface EnsPluginConfigs {
   query: QueryConfig;