+++ packages/js/client/src/pluginConfigs/Ethereum.ts
@@ -1,12 +1,12 @@
 /// NOTE: This is an auto-generated file. See scripts/extractPluginConfigs.ts
 /* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable prettier/prettier */
 
 /// Types generated from @web3api/ethereum-plugin-js build files:
 /// build/index.d.ts, build/common/EthereumConfig.d.ts, build/common/Connection.d.ts
 
-export interface EthereumPluginConfigs
-  extends EthereumConfig,
-    Record<string, unknown> {}
+export interface EthereumPluginConfigs extends EthereumConfig, Record<string, unknown> {
+}
 
 export interface EthereumConfig {
   networks: ConnectionConfigs;