+++ packages/js/plugins/http/src/__tests__/e2e/integration.spec.ts
@@ -1,5 +1,5 @@
 import { httpPlugin } from "../..";
-import { Response } from "../../query/w3-man";
+import { Response } from "../../query/w3";
 
 import { Web3ApiClient, defaultIpfsProviders } from "@web3api/client-js"
 import { ensPlugin } from "@web3api/ens-plugin-js";
@@ -33,6 +33,9 @@ describe("e2e tests for HttpPlugin", () => {
       const { data } = await axios.get("http://localhost:4040/deploy-ens");
 
       ensAddress = data.ensAddress
+      
+      const registrarAddress = data.registrarAddress
+      const resolverAddress = data.resolverAddress
 
       client = new Web3ApiClient({
         plugins: [
@@ -71,11 +74,14 @@ describe("e2e tests for HttpPlugin", () => {
         ],
       });
 
-      const api = await buildAndDeployApi(
-        `${__dirname}/integration`,
-        ipfs,
-        ensAddress
-      );
+      const api = await buildAndDeployApi({
+        apiAbsPath: `${__dirname}/integration`,
+        ipfsProvider: ipfs,
+        ensRegistryAddress: ensAddress,
+        ensRegistrarAddress: registrarAddress,
+        ensResolverAddress: resolverAddress,
+        ethereumProvider: ethereum
+      });
 
       uri = `ens/testnet/${api.ensDomain}`;
     });