+++ packages/js/plugins/http/src/__tests__/unit/util.test.ts
@@ -1,5 +1,5 @@
 import { fromAxiosResponse, toAxiosRequestConfig } from "../../query/util";
-import { ResponseTypeEnum } from "../../query/w3-man";
+import { ResponseTypeEnum } from "../../query/w3";
 
 describe("converting axios response", () => {
   test("response type: text", () => {