+++ packages/js/plugins/logger/src/query/index.ts
@@ -1,9 +1,4 @@
-import {
-  Module,
-  Input_log,
-  Logger_LogLevel,
-  Logger_LogLevelEnum,
-} from "./w3-man";
+import { Module, Input_log, Logger_LogLevel, Logger_LogLevelEnum } from "./w3";
 
 export type LogFunc = (level: Logger_LogLevel, message: string) => boolean;