+++ packages/test-cases/cases/bind/sanity/output/wasm-as/common/imported/TestImport_AnotherObject/index.ts
@@ -4,7 +4,9 @@ import {
   Nullable,
   BigInt,
   BigNumber,
-  JSON
+  JSON,
+  JSONSerializer,
+  JSONDeserializer,
 } from "@web3api/wasm-as";
 import {
   serializeTestImport_AnotherObject,
@@ -14,6 +16,7 @@ import {
 } from "./serialization";
 import * as Types from "../..";
 
+@serializable
 export class TestImport_AnotherObject {
 
   public static uri: string = "testimport.uri.eth";
@@ -34,5 +37,13 @@ export class TestImport_AnotherObject {
 
   static read(reader: Read): TestImport_AnotherObject {
     return readTestImport_AnotherObject(reader);
+  }
+
+  static toJson(type: TestImport_AnotherObject): JSON.Value {
+    return JSONSerializer.encode(type);
+  }
+
+  static fromJson(json: JSON.Value): TestImport_AnotherObject {
+    return (new JSONDeserializer(json)).decode<TestImport_AnotherObject>();
   }
 }